connector_builder_agents/pyproject.toml
[project]
name = "airbyte-connector-builder-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.1.1",
    "openai-agents-mcp>=0.0.8",
    "openai-agents>=0.2.11",
    "mcp-agent>=0.1.15",  # Transitive dependency of openai-agents
    "pydantic>=2.11.9,<3.0",
    "pyyaml>=6.0.1",
    "arize-phoenix>=5.0.0",
    "openinference-instrumentation-openai-agents",
    "pandas>=2.0.0",
]

[dependency-groups]
dev = [
    "poethepoet>=0.37.0",
]
